Added Except Pattern
This commit adds the Except pattern (and maybe the Ignorable pattern)
The Except pattern allows you to specifiy sets of patterns that match
some pattern and do not match another pattern. For example, many
programming languages have reserved keywords. Variable names cannot be
these keywords, so the Except pattern provides a way to exclude the set
of all characters that are keywords from the set of all words. Before
this pattern, you would have had to define a word as the set of all
characters that are not a keyword. By hand. Which would suck.

Use Except as pl.Except(pattern, not_pattern);